Add `include_file_descriptor_set` to `prost-build`
This commit adds support for writing an encoded version of the
appropriate FileDescriptorSet to a file named `file_descriptor_set.bin`
in the output directory. This can be used with the `include_bytes` macro
by consuming applications or libraries.

The rationale for this pull request is to support the gRPC Server
Reflection Protocol, which operates in terms of FileDescriptorProto, in
Tonic. FileDescriptorProto instances are contained inside a
FileDescriptorSet.